clis/monarch-cli/pyproject.toml
[project]
name = "gfd-monarch-cli"
version = "0.1.0"
requires-python = ">=3.12"
description = "Personal finance CLI wrapping Monarch Money"
dependencies = [
    "click>=8.0",
    "monarchmoney-enhanced>=0.11.0",
    "gql[aiohttp]>=3.4,<4",
    "keyring>=25.0",
    "tabulate>=0.9",
]

[project.scripts]
gfd-monarch = "gfd_monarch_cli.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gfd_monarch_cli"]
